Split general security configuration and crowd-specific configuration into seperate classes.
